+++ internal/domain/power-cap.go
@@ -514,7 +514,6 @@ func doPowerCapTask(taskID uuid.UUID) {
 			}
 		}
 		tempOps := []model.PowerCapOperation{}
-//		if comp.PowerCapControlsCount > 0 && !taskIsPatch {
 		if comp.PowerCapControlsCount > 0 {
 			// When a component is using the Controls schema, it is because each available power control
 			// is located at a different URL. Make an operation for each URL.
@@ -539,29 +538,18 @@ func doPowerCapTask(taskID uuid.UUID) {
 					// loop only to pick up the first pwrCtl.Path to form the /Controls.Deep URL
 					op.PowerCaps = comp.PowerCaps
 					tempOps = append(tempOps, op)
-logger.Log.Infof("JW_DEBUG ----------> CNTRLS PATCH: op.PowerCapURI=%s", op.PowerCapURI)
 					break
 				}
 
 				op.PowerCaps = make(map[string]hsm.PowerCap)
 				op.PowerCaps[name] = pwrCtl
 				tempOps = append(tempOps, op)
-logger.Log.Infof("JW_DEBUG ----------> CNTRLS NOT A PATCH: op.PowerCapURI=%s", op.PowerCapURI)
 			}
 		} else {
 			op := model.NewPowerCapOperation(task.TaskID, task.Type)
-//			if comp.PowerCapControlsCount > 0 {
-//				// Use Controls.Deep URL for Cray EX hardware.
-//				pwrURL := comp.PowerCapURI
-//				url := path.Dir(pwrURL)
-//				op.PowerCapURI = url + "/Controls.Deep"
-//			} else {
-//				op.PowerCapURI = comp.PowerCapURI
-//			}
 			op.PowerCapURI = comp.PowerCapURI
 			op.PowerCaps = comp.PowerCaps
 			tempOps = append(tempOps, op)
-logger.Log.Infof("JW_DEBUG ----------> NOT CNTRLS: op.PowerCapURI=%s", op.PowerCapURI)
 		}
 
 		// Validate that we have the required HSM data for each operation.
@@ -687,7 +675,6 @@ logger.Log.Infof("JW_DEBUG ----------> NOT CNTRLS: op.PowerCapURI=%s", op.PowerC
 				method = "PATCH"
 				path = op.RfFQDN + op.PowerCapURI
 			}
-logger.Log.Infof("JW_DEBUG ----------> doPowerCapTask: method=%s path=https://%s", method, path)
 			payload, _ := generatePowerCapPayload(op)
 			trsTaskList[trsTaskIdx].Request, _ = http.NewRequest(method, "https://"+path, bytes.NewBuffer(payload))
 			trsTaskList[trsTaskIdx].Request.Header.Set("Content-Type", "application/json")
@@ -778,7 +765,6 @@ logger.Log.Infof("JW_DEBUG ----------> doPowerCapTask: method=%s path=https://%s
 // Generates a POST/PATCH payload for the specified operation
 func generatePowerCapPayload(op model.PowerCapOperation) ([]byte, error) {
 	if op.PowerCapControlsCount > 0 {
-logger.Log.Infof("JW_DEBUG ----------> generatePowerCapPayload: CNTRLS: op.PowerCapControlsCount=%d", op.PowerCapControlsCount)
 		// Newer bard peak power capping schema deep patch
 		p := RFControlsDeep{
 			Members: make([]RFControl, 0, 1),
@@ -798,11 +784,9 @@ logger.Log.Infof("JW_DEBUG ----------> generatePowerCapPayload: CNTRLS: op.Power
 				}
 			}
 			p.Members = append(p.Members, ctl)
-logger.Log.Infof("JW_DEBUG ----------> generatePowerCapPayload: CNTRLS: limit.CurrentValue=%d", *limit.CurrentValue)
 		}
 		return json.Marshal(p)
 	} else if isHpeApollo6500(op.PowerCapURI) {
-logger.Log.Infof("JW_DEBUG ----------> generatePowerCapPayload: is Apollo 6500")
 		// Apollo 6500 redfish POST power capping schema
 		zero := 0
 		p := HpeConfigurePowerLimit{
@@ -815,7 +799,6 @@ logger.Log.Infof("JW_DEBUG ----------> generatePowerCapPayload: is Apollo 6500")
 		}
 		return json.Marshal(p)
 	} else {
-logger.Log.Infof("JW_DEBUG ----------> generatePowerCapPayload: is generic: op.PowerCapCtlInfoCount=%d", op.PowerCapCtlInfoCount)
 		// Generic redfish power capping schema
 		p := Power{
 			PowerCtl: make([]PowerControl, op.PowerCapCtlInfoCount),
@@ -827,7 +810,6 @@ logger.Log.Infof("JW_DEBUG ----------> generatePowerCapPayload: is generic: op.P
 					LimitInWatts: limit.CurrentValue,
 				},
 			}
-logger.Log.Infof("JW_DEBUG ----------> generatePowerCapPayload: is generic: limit.CurrentValue=%d", *limit.CurrentValue)
 			p.PowerCtl[idx] = pCtl
 		}
 		return json.Marshal(p)